--- a/justfile
+++ b/justfile
@@ -46,7 +46,7 @@ build:
     if [[ "${RUST_APT_WORKER_SIZES:+x}" == 'x' ]]; then
         extra_cargo_args+=('--features' 'rust-apt/worker_sizes')
     fi
-	
+
     cargo build --target "${target}" --release "${extra_cargo_args[@]}"
     install -Dm 755 "target/${target}/release/makedeb-rs" target/release/makedeb-rs