--- a/test/prepare/rust.bats
+++ b/test/prepare/rust.bats
@@ -1,7 +1,0 @@
-@test "make sure rust code is formatted" {
-    cargo fmt --check
-}
-
-@test "make sure cargo clippy lints pass" {
-    cargo clippy -- -D warnings
-}