--- a/test/tests/arch.bats
+++ b/test/tests/arch.bats
@@ -1,5 +1,6 @@
 load ../util/util
 
+# bats test_tags=lint
 @test "correct arch - all allowed characters" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -8,9 +9,10 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch x86_64
     pkgbuild clean
-    makedeb -d
+    makedeb --lint
 }
 
+# bats test_tags=lint
 @test "incorrect arch - use any with another architecture" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -19,11 +21,11 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any x86_64
     pkgbuild clean
-    run makedeb -d
-    [[ "${status}" == "12" ]]
+    run -12 makedeb --lint
     [[ "${output}" == "[!] Can not use 'any' architecture with other architectures" ]]
 }
 
+# bats test_tags=lint
 @test "incorrect arch - unknown architecture" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -32,7 +34,6 @@ load ../util/util
     pkgbuild string pkgrel 1
     pkgbuild array arch there_is_no_way_this_architecture_exists
     pkgbuild clean
-    run makedeb -d
-    [[ "${status}" == "12" ]]
+    run -12 makedeb --lint
     [[ "${output}" == "[!] testpkg is not available for the 'amd64' architecture." ]]
 }