--- a/test/tests/conflicts.bats
+++ b/test/tests/conflicts.bats
@@ -11,6 +11,6 @@ load ../util/util
     pkgbuild clean
     makedeb -s --install --no-confirm --allow-downgrades
     [[ "$(cat pkg/testpkg/DEBIAN/control | grep '^Conflicts')" == "Conflicts: krita (>> 0), asciidoctor" ]]
-    run ! type make
     run ! type asciidoctor
+    run ! type krita
 }