--- a/test/tests/functions.bats
+++ b/test/tests/functions.bats
@@ -1,5 +1,6 @@
 load ../util/util
 
+# bats test_tags=lint
 @test "correct prepare()" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -8,7 +9,7 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    makedeb -d
+    makedeb --lint
 }
 
 @test "incorrect prepare() - bad exit code" {
@@ -21,10 +22,11 @@ load ../util/util
     remove_function "prepare"
     echo -e '\nprepare() { false; }' >> PKGBUILD
     pkgbuild clean
-    run makedeb -d
-    [[ "${status}" == "4" ]]
+    run -4 makedeb -d
+    # [[ "${status}" == "4" ]]
 }
 
+# bats test_tags=lint
 @test "incorrect pkgver() - incorrect syntax for returned version" {
     pkgver() {
         echo "hi me"
@@ -38,10 +40,11 @@ load ../util/util
     pkgbuild array arch any
     pkgbuild function pkgver
     pkgbuild clean
-    run makedeb -d
-    [[ "${status}" == "12" ]]
+    run -12 makedeb -d
+    # [[ "${status}" == "12" ]]
 }
 
+# bats test_tags=lint
 @test "incorrect package() - missing 'package()' function" {
     build() {
         true
@@ -55,6 +58,6 @@ load ../util/util
     pkgbuild array arch any
     pkgbuild function build
     pkgbuild clean
-    run makedeb -d
-    [[ "${status}" == "12" ]]
+    run -12 makedeb --lint
+    # [[ "${status}" == "12" ]]
 }