--- a/test/tests/maintainer.bats
+++ b/test/tests/maintainer.bats
@@ -1,5 +1,6 @@
 load ../util/util
 
+# bats test_tags=lint
 @test "correct maintainer" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -11,6 +12,7 @@ load ../util/util
     makedeb --lint
 }
 
+# bats test_tags=lint
 @test "incorrect maintainer - no maintainer" {
     pkgbuild string pkgname testpkg
     pkgbuild string pkgver 1.0.0
@@ -22,6 +24,7 @@ load ../util/util
     [[ "${output}" == "[!] A maintainer must be specified. This will be an error in a future release." ]]
 }
 
+# bats test_tags=lint
 @test "incorrect maintainer - more than one maintainer" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string maintainer2 'Bar Foo <bar@foo.com>'