--- a/test/tests/misc.bats
+++ b/test/tests/misc.bats
@@ -11,6 +11,7 @@ load ../util/util
     makedeb -si --no-confirm --allow-downgrades
 }
 
+# bats test_tags=lint
 @test "correct - set dependencies from a distro-dependency variable" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -34,6 +35,7 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
+    # FIXME Check command output with generated files
     makedeb --print-srcinfo
     makedeb --print-control
 }