--- a/test/tests/optdepends.bats
+++ b/test/tests/optdepends.bats
@@ -1,5 +1,6 @@
 load ../util/util
 
+# bats test_tags=lint
 @test "correct optdepends - all valid characters" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -9,7 +10,7 @@ load ../util/util
     pkgbuild array arch any
     pkgbuild array optdepends 'bats>0: good shell testing framework' 'bash: king of shell'
     pkgbuild clean
-    makedeb -d
+    makedeb --lint
 }
 
 @test "correct optdepends - install missing dependencies" {
@@ -40,6 +41,7 @@ load ../util/util
     [[ "$(cat pkg/testpkg/DEBIAN/control | grep 'Recommends:')" == "Recommends: bats (>> 0)" ]]
 }
 
+# bats test_tags=lint
 @test "incorrect optdepends - invalid dependency prefix" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -49,7 +51,7 @@ load ../util/util
     pkgbuild array arch all
     pkgbuild array optdepends 'z!bats'
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
     [[ "${status}" == "12" ]]
     [[ "${output}" == "[!] Dependency 'z!bats' under 'optdepends' contains an invalid prefix: 'z'" ]]
 }