--- a/test/tests/optdepends.bats
+++ b/test/tests/optdepends.bats
@@ -21,11 +21,12 @@ load ../util/util
     pkgbuild string pkgrel 1
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
-    pkgbuild array optdepends 'zsh: adding a reason so we can see if makedeb strips it before adding it to the control file' 'yash>=0.0.1'
+    pkgbuild array optdepends 'sox: adding a reason so we can see if makedeb strips it before adding it to the control file' 'yash>=0.0.1'
     pkgbuild clean
     makedeb -s --no-confirm --allow-downgrades
 }
 
+# bats test_tags=lint
 @test "correct optdepends - valid dependency prefixes" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgname testpkg
@@ -35,10 +36,10 @@ load ../util/util
     pkgbuild array arch any
     pkgbuild array optdepends 'r!bats>0' 's!bash' 'yash'
     pkgbuild clean
-    makedeb -d
+    run makedeb --print-control
 
-    [[ "$(cat pkg/testpkg/DEBIAN/control | grep 'Suggests:')" == "Suggests: bash, yash" ]]
-    [[ "$(cat pkg/testpkg/DEBIAN/control | grep 'Recommends:')" == "Recommends: bats (>> 0)" ]]
+    [[ "$(echo "$output" | grep 'Suggests:')" == "Suggests: bash, yash" ]]
+    [[ "$(echo "$output" | grep 'Recommends:')" == "Recommends: bats (>> 0)" ]]
 }
 
 # bats test_tags=lint