--- a/test/tests/pkgbase.bats
+++ b/test/tests/pkgbase.bats
@@ -1,5 +1,6 @@
 load ../util/util
 
+# bats test_tags=lint
 @test "correct pkgbase - lowercase letters" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgbase testpkg
@@ -9,9 +10,10 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    makedeb -d
+    makedeb --lint
 }
 
+# bats test_tags=lint
 @test "correct pkgbase - digits and minus sign" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgbase test-123
@@ -21,9 +23,10 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
 }
 
+# bats test_tags=lint
 @test "correct pkgbase - plus sign" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgbase 'test-1+3'
@@ -33,9 +36,10 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
 }
 
+# bats test_tags=lint
 @test "correct pkgbase - period" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgbase 'test-1.3'
@@ -45,9 +49,10 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
 }
 
+# bats test_tags=lint
 @test "incorrect pkgbase - capital letters" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgbase test-Akg
@@ -57,11 +62,12 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
     [[ "${status}" == "12" ]]
     [[ "${output}" == "[!] 'pkgbase' contains capital letters" ]]
 }
 
+# bats test_tags=lint
 @test "incorrect pkgbase - invalid character" {
     pkgbuild string maintainer1 'Foo Bar <foo@bar.com>'
     pkgbuild string pkgbase 'test-#kg'
@@ -71,7 +77,7 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
     [[ "${status}" == "12" ]]
     [[ "${output}" == "[!] pkgbase contains invalid characters: '#'" ]]
 }
@@ -85,7 +91,7 @@ load ../util/util
     pkgbuild string pkgdesc "package description"
     pkgbuild array arch any
     pkgbuild clean
-    run makedeb -d
+    run makedeb --lint
     [[ "${status}" == "12" ]]
     [[ "${output}" == '[!] pkgbase should not be an array' ]]
 }