… test/prepare/prepare.bats
… test/tests/pkgdesc.bats
… test/tests/pkgname.bats
… test/tests/pkgrel.bats
… test/tests/pkgver.bats
… test/tests/provides.bats
… test/tests/source.bats
… test/tests/vcs.bats
… test/util/util.bash
… toast.yml