The purpose of this PR is to enable faster running of the unit tests.

## Changes:

- Use the latest `bats-core`
- Run `makedeb --lint` tests in parallel
- Refactor `toast.yml` to only rebuild when the relevant files have changed

## Results:

### Base line

1. Full build

```sh
Executed in   15.54 mins    fish           external
   usr time    1.33 secs  710.00 micros    1.33 secs
   sys time    1.14 secs  242.00 micros    1.14 secs
```

2. Partial build - Only changed tests

```sh
Executed in  573.08 secs      fish           external
   usr time  611.11 millis  953.00 micros  610.16 millis
   sys time  438.46 millis  337.00 micros  438.12 millis
```

### Current

1. Full build

```sh
Executed in  630.34 secs    fish           external
   usr time    3.87 secs    1.25 millis    3.87 secs
   sys time    2.51 secs    0.33 millis    2.51 secs
```

2. Partial Build - Only changed tests

```sh
Executed in  116.94 secs      fish           external
   usr time  335.65 millis  594.00 micros  335.06 millis
   sys time  311.76 millis  214.00 micros  311.55 millis
```

## TODO

- [ ] Integrate `toast.yml` - `run_unit_tests` commands with `drone` scripts.