@hwittenborn 2023-06-06
How would the tests get triggered now? The part that's failing is from here: https://github.com/makedeb/makedeb/blob/alpha/.drone/scripts/run-unit-tests.sh#L9.