@bithium 2023-06-06
Hi, I run the tests using `toast run_unit_tests`. The `prepare` portion of the tests is now done in the `build_makedeb-rs` .

That is part that is missing with the integration with `drone`. 
Can we simply run the `toast` command ? 
That way the workflow would  be the same as both in local development and in CI.